add resetToFirstGeneration() in GameOfLiveManager
with this method it is possible to reset the game of life to the first generation,
so the cells from the first generation are brought back to life (if they are dead)
and all other living cells get killed.
Returns the coordinates of the cells which got a new life status (dead or alive).