add method for toggling buttons in fieldPanel
With toggleButtonsInFieldPanel it is possible for the resetPanel to
toggle buttons at certain positions in the fieldPanel.
This can be used if there are changes in the life status of cells outside of the
timer interval e.g. when resetting the game to its first generation or
killig all cells.