make GameOfLifeField an Observable
Now GameOfLifeField notifies its observers if data has changed.
The observers can retreive data from the model (GameOfLifeField) to update their UI elements.
For retreiving the positions of cells which got a new life state, the GameOflifeField
has a list which contains all cells with a new life state.
Every time an observable asscesses this list (with clearCellsToBeUpdated), it will be emptied.