create new GameOfLife class
this new GameOfLife class is the main model object which implements the observer interface.
It contains a gameOfLifeField object, methods to interact with the field and more functionalities
(like generationCounter) that are needed for the game of life.
The goal was to move all methods and fields which are not directly related to a
gameOfLife field into a new class to follow the oop idea.