implement and use goTo feature
This feature makes it possible to go to a certain
generation (after or before the current one).
If the requested generation comes before the current one,
the requested generation will be loaded starting with the first generation,
so manual changes to cells life state or color after the first generation are ignored.